backend/requirements.txt
beautifulsoup4==4.10.0
datasets==3.5.0
evaluate==0.4.3
Flask==3.1.1
flask_cors==6.0.1
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.24.4
pandas==2.3.0
Requests==2.32.4
serpapi==0.1.5
torch==2.2.0
transformers==4.30.2
